use symbol type as enum
